requirements.txt
# Vaani — Python dependencies

# ── Core ──────────────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Cache / Queue ─────────────────────────────────────────────────────────────
redis[asyncio]==5.2.0
celery==5.4.0
celery[redis]==5.4.0
croniter==3.0.3

# ── AI ────────────────────────────────────────────────────────────────────────
anthropic==0.40.0
openai==1.57.2

# ── HTTP ─────────────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Auth / Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.18

# ── Messaging ─────────────────────────────────────────────────────────────────
python-telegram-bot==21.9

# ── Payments ──────────────────────────────────────────────────────────────────
razorpay==1.4.2

# ── PDF Generation ────────────────────────────────────────────────────────────
reportlab==4.2.5

# ── AWS ───────────────────────────────────────────────────────────────────────
boto3==1.35.81

# ── Logging ──────────────────────────────────────────────────────────────────
structlog==24.4.0

# ── Dev ───────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1  # for test client
